requirements.txt
requests==2.18.4
pymongo==3.6.1
html2text==2018.1.9
jieba==0.39
beautifulsoup4==4.6.0
readability_lxml==0.6.2
scikit_learn==0.19.1
